# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: LendingOS — Multi-Tenant AI Lending Operating System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /collections:
    get:
      operationId: listCollections
      tags: [collections]
      summary: List collection cases
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [new, in_progress, resolved, escalated]
        - name: priority
          in: query
          schema:
            type: string
            enum: [low, medium, high, critical]
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
      responses:
        "200":
          description: Collection cases
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CollectionListResponse"
  /collections/{collectionId}:
    get:
      operationId: getCollection
      tags: [collections]
      summary: Get collection case
      parameters:
        - name: collectionId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Collection case
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Collection"
    patch:
      operationId: updateCollection
      tags: [collections]
      summary: Update collection case
      parameters:
        - name: collectionId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CollectionUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Collection"

  # ─── ANALYTICS ──────────────────────────────────────────────────────────────
...
  /analytics/tenant-dashboard:
    get:
      operationId: getTenantDashboard
      tags: [analytics]
      summary: Tenant dashboard metrics
      parameters:
        - name: tenantId
          in: query
          schema: { type: string }
        - name: period
          in: query
          schema: { type: string, enum: [7d, 30d, 90d, 1y], default: "30d" }
      responses:
        "200":
          description: Tenant dashboard
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantDashboard"
  /analytics/loan-funnel:
    get:
      operationId: getLoanFunnel
      tags: [analytics]
      summary: Loan application funnel analytics
      parameters:
        - name: tenantId
          in: query
          schema: { type: string }
        - name: period
          in: query
          schema: { type: string, enum: [7d, 30d, 90d, 1y], default: "30d" }
      responses:
        "200":
          description: Funnel data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoanFunnel"
  /analytics/collection-rate:
    get:
      operationId: getCollectionRate
      tags: [analytics]
      summary: Collection rate over time
      parameters:
        - name: tenantId
          in: query
          schema: { type: string }
        - name: period
          in: query
          schema: { type: string, enum: [7d, 30d, 90d, 1y], default: "30d" }
      responses:
        "200":
          description: Collection rate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CollectionRateData"
  /analytics/revenue-trend:
    get:
      operationId: getRevenueTrend
      tags: [analytics]
      summary: Revenue trend data
      parameters:
        - name: tenantId
          in: query
          schema: { type: string }
        - name: period
          in: query
          schema: { type: string, enum: [7d, 30d, 90d, 1y], default: "30d" }
      responses:
        "200":
          description: Revenue trend
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RevenueTrend"

  # ─── AUDIT LOGS ─────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Collection:
      type: object
      required: [id, loanId, customerId, overdueAmount, dpd, status, priority]
      properties:
        id: { type: string }
        loanId: { type: string }
        loanNumber: { type: string }
        customerId: { type: string }
        customerName: { type: string }
        customerPhone: { type: string }
        overdueAmount: { type: number }
        dpd: { type: integer }
        status:
          type: string
          enum: [new, in_progress, resolved, escalated]
        priority:
          type: string
          enum: [low, medium, high, critical]
        aiPriorityScore: { type: ["number", "null"] }
        assignedTo: { type: ["string", "null"] }
        lastContactAt: { type: ["string", "null"] }
        nextFollowUpAt: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
        createdAt: { type: string }
    CollectionUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [new, in_progress, resolved, escalated]
        priority:
          type: string
          enum: [low, medium, high, critical]
        assignedTo: { type: string }
        nextFollowUpAt: { type: string }
        notes: { type: string }
    CollectionListResponse:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Collection"
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }

    # ── ANALYTICS ─────────────────────────────────────────────────────────────
...
    TenantDashboard:
      type: object
      required: [period, applications, disbursals, collections, revenue, customers]
      properties:
        period: { type: string }
        applications:
          type: object
          required: [total, approved, rejected, pending]
          properties:
            total: { type: integer }
            approved: { type: integer }
            rejected: { type: integer }
            pending: { type: integer }
        disbursals:
          type: object
          required: [total, amount]
          properties:
            total: { type: integer }
            amount: { type: number }
        collections:
          type: object
          required: [collected, overdue, rate]
          properties:
            collected: { type: number }
            overdue: { type: number }
            rate: { type: number }
        revenue:
          type: object
          required: [interest, fees, total]
          properties:
            interest: { type: number }
            fees: { type: number }
            total: { type: number }
        customers:
          type: object
          required: [new, active, total]
          properties:
            new: { type: integer }
            active: { type: integer }
            total: { type: integer }
        trend:
          type: array
          items:
            type: object
            required: [date, disbursals, collections]
            properties:
              date: { type: string }
              disbursals: { type: number }
              collections: { type: number }
    LoanFunnel:
      type: object
      required: [period, stages]
      properties:
        period: { type: string }
        stages:
          type: array
          items:
            type: object
            required: [stage, count, dropoff]
            properties:
              stage: { type: string }
              count: { type: integer }
              dropoff: { type: number }
    CollectionRateData:
      type: object
      required: [period, overallRate, trend]
      properties:
        period: { type: string }
        overallRate: { type: number }
        trend:
          type: array
          items:
            type: object
            required: [date, rate, collected, overdue]
            properties:
              date: { type: string }
              rate: { type: number }
              collected: { type: number }
              overdue: { type: number }
    RevenueTrend:
      type: object
      required: [period, total, trend]
      properties:
        period: { type: string }
        total: { type: number }
        trend:
          type: array
          items:
            type: object
            required: [date, interest, fees, total]
            properties:
              date: { type: string }
              interest: { type: number }
              fees: { type: number }
              total: { type: number }

    # ── AUDIT LOGS ────────────────────────────────────────────────────────────
...